+++ package.json
@@ -3,23 +3,20 @@
   "version": "1.0.1",
   "description": "Fetch keys for object with a default or fail",
   "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/kaspernj/object-fetch.git"
+    "url": "git+ssh://git@github.com/kaspernj/object-digger.git"
   },
   "author": "kaspernj",
   "license": "ISC",
   "bugs": {
-    "url": "https://github.com/kaspernj/object-fetch/issues"
+    "url": "https://github.com/kaspernj/object-digger/issues"
   },
-  "homepage": "https://github.com/kaspernj/object-fetch#readme",
+  "homepage": "https://github.com/kaspernj/object-digger#readme",
   "devDependencies": {
-    "jest": "^26.5.3"
+    "jasmine": "^3.6.3"
   },
   "scripts": {
-    "test": "jest"
+    "test": "jasmine"
   }
 }
